Fix build and update dependencies to latest
This PR updates all all `dependencies` and `devDependencies` to the
latest versions and updates tests so they build with the latest
`devDependencies`.

There is a breaking change to update `engines` to `>=7.6.0` due to an
`xo` lint. Exported functions are now `async` functions.

`npm test` exits with a non-zero exit code from `ava` that I'm not sure
how to address.